+++ ext_tables.php
@@ -11,9 +11,33 @@ call_user_func(static function() {
     \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::allowTableOnStandardPages('tx_qcinforights_domain_model_qcinforights');
 });
 
+/**
+ * Checking if the be_user is not null to prepare his data
+ */
+if($GLOBALS['BE_USER'] === null) {
+    $GLOBALS['BE_USER'] = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance(\TYPO3\CMS\Core\Authentication\BackendUserAuthentication::class);
+    $GLOBALS['BE_USER']->start();
+    $GLOBALS['BE_USER']->fetchGroupData();
+}
+
+//Render user TsConfig
+$userTS = $GLOBALS['BE_USER']->getTSConfig()['mod.']['qcinforights.'];
+
+//Rendere Page TsConfig by default get first page
 $modTSconfig = BackendUtility::getPagesTSconfig(1)['mod.']['qcinforights.'];
 
-if($modTSconfig['showTabAccess'] == 1 || $modTSconfig['showMenuAccess']) {
+//Checking about access
+$showMenuAccess =  (int)checkShowTsConfig($userTS,$modTSconfig,'showMenuAccess');
+$showMenuGroups =  (int)checkShowTsConfig($userTS,$modTSconfig,'showMenuGroups');
+$showMenuUsers  =  (int)checkShowTsConfig($userTS,$modTSconfig,'showMenuUsers');
+
+
+//@deprecated will removed in the next update v1.3.0
+$showTabAccess =   (int)checkShowTsConfig($userTS,$modTSconfig,'showTabAccess');
+$showTabGroups =   (int)checkShowTsConfig($userTS,$modTSconfig,'showTabGroups');
+$showTabUsers  =    (int)checkShowTsConfig($userTS,$modTSconfig,'showTabUsers');
+
+if($showMenuAccess || $showTabAccess) {
     // Extend Module INFO with new Element for access and rights tab
     \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::insertModuleFunction(
         'web_info',
@@ -24,7 +48,7 @@ if($modTSconfig['showTabAccess'] == 1 || $modTSconfig['showMenuAccess']) {
 }
 
 // Extend Module INFO for Groups tab
-if($modTSconfig['showTabGroups'] == 1 || $modTSconfig['showMenuGroups'] == 1) {
+if($showTabGroups || $showMenuGroups) {
     \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::insertModuleFunction(
         'web_info',
         GroupsReport::class,
@@ -33,8 +57,8 @@ if($modTSconfig['showTabGroups'] == 1 || $modTSconfig['showMenuGroups'] == 1) {
     );
 }
 
-if($modTSconfig['showTabUsers'] == 1 || $modTSconfig['showMenuUsers'] == 1){
 // Extend Module INFO For Users tab
+if($showTabUsers || $showMenuUsers){
     \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::insertModuleFunction(
         'web_info',
         UsersReport::class,
@@ -48,3 +72,21 @@ if($modTSconfig['showTabUsers'] == 1 || $modTSconfig['showMenuUsers'] == 1){
     'qcinforights',
     'EXT:qc_info_rights/Resources/Private/Language/Module/locallang_csh.xlf'
 );
+
+/**
+ * PHP function to check and validate the access&right for each mo
+ * @param array  $userTS
+ * @param array  $modTSconfig
+ * @param string $value
+ *
+ * @return string
+ */
+function checkShowTsConfig(array $userTS, array $modTSconfig, string $value): string
+{
+    if (is_array($userTS) && array_key_exists($value, $userTS)) {
+        return $userTS[$value];
+    } else if (is_array($modTSconfig) && array_key_exists($value, $modTSconfig)) {
+        return $modTSconfig[$value];
+    }
+    return '';
+}