+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Info Rights
+## Change log v1.2.1
+### Bug Fix
+
+- 2022-05-18 [BUGFIX] Add missing user TsConfig to verify access to the context menu item 
+### Breaking
+
+- 2022-05-16 [!!!][TASK]  Replace TsConfig showTabAccess,showTabUsers,showTabGroups with the new showMenuAccess,showMenuUsers, and showMenuGroups. The old TSconfig keyes will be removed in next update v 1.3.0
+
 ## Change log  v1.2.0
-2022-05-16 [Feature] We replace the 2 tabs (user list and group list) with menu items.    
-2022-05-16 [!!!][TASK]  Replace TsConfig showTabAccess,showTabUsers,showTabGroups with the new showMenuAccess,showMenuUsers, and showMenuGroups. The old TSconfig keyes will be removed in next update
+- 2022-05-16 [Feature] We replace the 2 tabs (user list and group list) with menu items.    
+- 2022-05-16 [!!!][TASK]  Replace TsConfig showTabAccess,showTabUsers,showTabGroups with the new showMenuAccess,showMenuUsers, and showMenuGroups. The old TSconfig keyes will be removed in next update
 ## Change log  v1.1.0
-2022-01-13 [Feature] It is now possible to get the members of groups when checking the pages rights
+- 2022-01-13 [Feature] It is now possible to get the members of groups when checking the pages rights